fix template build issue
