+++ monitoring/statistics.cc
@@ -3,12 +3,9 @@
 //  COPYING file in the root directory) and Apache 2.0 License
 //  (found in the LICENSE.Apache file in the root directory).
 //
+
 #include "monitoring/statistics.h"
 
-#include <cinttypes>
-#include "port/likely.h"
-#include <algorithm>
-#include <cstdio>
 
 namespace rocksdb {
 
@@ -227,206 +224,5 @@ const std::vector<std::pair<Histograms, std::string>> HistogramsNameMap = {
     {SST_BATCH_SIZE, "rocksdb.sst.batch.size"},
 };
 
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-std::shared_ptr<Statistics> CreateDBStatistics() {
-  return std::make_shared<StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>>(nullptr);
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::StatisticsImpl(
-    std::shared_ptr<Statistics> stats)
-    : stats_(std::move(stats)) {}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::~StatisticsImpl() {}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-uint64_t StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::getTickerCount(
-    uint32_t tickerType) const {
-  MutexLock lock(&aggregate_lock_);
-  return getTickerCountLocked(tickerType);
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-uint64_t StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::getTickerCountLocked(
-    uint32_t tickerType) const {
-  assert(tickerType < TICKER_MAX);
-  uint64_t res = 0;
-  for (size_t core_idx = 0; core_idx < per_core_stats_.Size(); ++core_idx) {
-    res += per_core_stats_.AccessAtCore(core_idx)->tickers_[tickerType];
-  }
-  return res;
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-void StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::histogramData(
-    uint32_t histogramType, HistogramData* const data) const {
-  MutexLock lock(&aggregate_lock_);
-  getHistogramImplLocked(histogramType)->Data(data);
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-std::unique_ptr<HistogramImpl>
-StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::getHistogramImplLocked(
-    uint32_t histogramType) const {
-  assert(histogramType < HISTOGRAM_MAX);
-  std::unique_ptr<HistogramImpl> res_hist(new HistogramImpl());
-  for (size_t core_idx = 0; core_idx < per_core_stats_.Size(); ++core_idx) {
-    res_hist->Merge(
-        per_core_stats_.AccessAtCore(core_idx)->histograms_[histogramType]);
-  }
-  return res_hist;
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-std::string StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::getHistogramString(
-    uint32_t histogramType) const {
-  MutexLock lock(&aggregate_lock_);
-  return getHistogramImplLocked(histogramType)->ToString();
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-void StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::setTickerCount(
-    uint32_t tickerType, uint64_t count) {
-  {
-    MutexLock lock(&aggregate_lock_);
-    setTickerCountLocked(tickerType, count);
-  }
-  if (stats_ && tickerType < TICKER_MAX) {
-    stats_->setTickerCount(tickerType, count);
-  }
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-void StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::setTickerCountLocked(
-    uint32_t tickerType, uint64_t count) {
-  assert(tickerType < TICKER_MAX);
-  for (size_t core_idx = 0; core_idx < per_core_stats_.Size(); ++core_idx) {
-    if (core_idx == 0) {
-      per_core_stats_.AccessAtCore(core_idx)->tickers_[tickerType] = count;
-    } else {
-      per_core_stats_.AccessAtCore(core_idx)->tickers_[tickerType] = 0;
-    }
-  }
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-uint64_t StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::getAndResetTickerCount(
-    uint32_t tickerType) {
-  uint64_t sum = 0;
-  {
-    MutexLock lock(&aggregate_lock_);
-    assert(tickerType < TICKER_MAX);
-    for (size_t core_idx = 0; core_idx < per_core_stats_.Size(); ++core_idx) {
-      sum +=
-          per_core_stats_.AccessAtCore(core_idx)->tickers_[tickerType].exchange(
-              0, std::memory_order_relaxed);
-    }
-  }
-  if (stats_ && tickerType < TICKER_MAX) {
-    stats_->setTickerCount(tickerType, 0);
-  }
-  return sum;
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-void StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::recordTick(uint32_t tickerType,
-                                                           uint64_t count) {
-  assert(tickerType < TICKER_MAX);
-  per_core_stats_.Access()->tickers_[tickerType].fetch_add(
-      count, std::memory_order_relaxed);
-  if (stats_ && tickerType < TICKER_MAX) {
-    stats_->recordTick(tickerType, count);
-  }
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-void StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::recordInHistogram(
-    uint32_t histogramType, uint64_t value) {
-  assert(histogramType < HISTOGRAM_MAX);
-  if (get_stats_level() <= StatsLevel::kExceptHistogramOrTimers) {
-    return;
-  }
-  per_core_stats_.Access()->histograms_[histogramType].Add(value);
-  if (stats_ && histogramType < HISTOGRAM_MAX) {
-    stats_->recordInHistogram(histogramType, value);
-  }
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-Status StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::Reset() {
-  MutexLock lock(&aggregate_lock_);
-  for (uint32_t i = 0; i < TICKER_MAX; ++i) {
-    setTickerCountLocked(i, 0);
-  }
-  for (uint32_t i = 0; i < HISTOGRAM_MAX; ++i) {
-    for (size_t core_idx = 0; core_idx < per_core_stats_.Size(); ++core_idx) {
-      per_core_stats_.AccessAtCore(core_idx)->histograms_[i].Clear();
-    }
-  }
-  return Status::OK();
-}
-
-namespace {
-
-// a buffer size used for temp string buffers
-const int kTmpStrBufferSize = 200;
-
-} // namespace
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-std::string StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::ToString() const {
-  MutexLock lock(&aggregate_lock_);
-  std::string res;
-  res.reserve(20000);
-  for (const auto& t : TickersNameMap) {
-    assert(t.first < TICKER_MAX);
-    char buffer[kTmpStrBufferSize];
-    snprintf(buffer, kTmpStrBufferSize, "%s COUNT : %" PRIu64 "\n",
-             t.second.c_str(), getTickerCountLocked(t.first));
-    res.append(buffer);
-  }
-  for (const auto& h : HistogramsNameMap) {
-    assert(h.first < HISTOGRAM_MAX);
-    char buffer[kTmpStrBufferSize];
-    HistogramData hData;
-    getHistogramImplLocked(h.first)->Data(&hData);
-    // don't handle failures - buffer should always be big enough and arguments
-    // should be provided correctly
-    int ret =
-        snprintf(buffer, kTmpStrBufferSize,
-                 "%s P50 : %f P95 : %f P99 : %f P100 : %f COUNT : %" PRIu64
-                 " SUM : %" PRIu64 "\n",
-                 h.second.c_str(), hData.median, hData.percentile95,
-                 hData.percentile99, hData.max, hData.count, hData.sum);
-    if (ret < 0 || ret >= kTmpStrBufferSize) {
-      assert(false);
-      continue;
-    }
-    res.append(buffer);
-  }
-  res.shrink_to_fit();
-  return res;
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-bool StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::getTickerMap(
-    std::map<std::string, uint64_t>* stats_map) const {
-  assert(stats_map);
-  if (!stats_map) return false;
-  stats_map->clear();
-  MutexLock lock(&aggregate_lock_);
-  for (const auto& t : TickersNameMap) {
-    assert(t.first < TICKER_MAX);
-    (*stats_map)[t.second.c_str()] = getTickerCountLocked(t.first);
-  }
-  return true;
-}
-
-template <uint32_t TICKER_MAX, uint32_t HISTOGRAM_MAX>
-bool StatisticsImpl<TICKER_MAX, HISTOGRAM_MAX>::HistEnabledForType(
-    uint32_t type) const {
-  return type < HISTOGRAM_MAX;
-}
 
 } // namespace rocksdb
